Fixed minor issue

+++ myfitnesstrack-api/src/main/java/com/myfitnesstrack/myfitnesstrackapi/progress/ProgressTable.java
@@ -31,7 +31,7 @@ public class ProgressTable {
 
     @Min(value = 0, message = "Calories taken must be a positive value")
     @Column(name = "calories_taken")
-    private double caloriesTaken;
+    private int caloriesTaken;
 
     @NotNull(message = "Date is required")
     private LocalDate date;
